Move test lab into end-to-end package
